// web/src/lib/api/generated/models/githubComNgodingvarengMemoriaInternalDeliveryRestDtoPublicUserResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Book API
 * OpenAPI spec version: 1.0
 */

export interface GithubComNgodingvarengMemoriaInternalDeliveryRestDtoPublicUserResponse {
  bio?: string;
  id?: string;
  image_path?: string;
  name?: string;
  /** Username is nil until the user has claimed one. */
  username?: string;
}